+++ gems/aws-sdk-quicksight/lib/aws-sdk-quicksight/client.rb
@@ -12529,7 +12529,7 @@ module Aws::QuickSight
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-quicksight'
-      context[:gem_version] = '1.89.0'
+      context[:gem_version] = '1.90.0'
       Seahorse::Client::Request.new(handlers, context)
     end